Respect the manifest's assemblyPrefix value on the server
